==> australia/programas/02_anomalias.ipynb <==
## 02 - Cáculo de la Anomalías de Temperatura
Se calculará para cada *borehole* una regresión lineal con los datos a partir de los 100m para así obtener el estado cuasi - estático. A continuación se calcular las anomalías de temperaturas en diferentes profundidades, restando los datos reales a los estimados con la recta.

# Librerías de procesado y calculo de datos
import numpy as np
import pandas as pd

# Librerías para hacer gráficos
import matplotlib.pyplot as plt
import seaborn as sns

# Librerias importantes
import warnings
from typing import Tuple
warnings.filterwarnings("ignore")

# Especifico el estilo que quiero en mis gráficos
sns.set_theme(style="whitegrid")

plt.ioff()

# Leo los datos obtenidos de los documentos y los recopilo en un dataframe.
path = "../boreholes/dataframe.csv"
df = pd.read_csv(path,delimiter=",",decimal=".",index_col="Unnamed: 0")
df.head(2)

#Obtengo en una lista donde comenzarían los datos en cada caso
posicion = df["Posición del Primer Dato"]
names = df.index

# Saco el path de cada uno de los archivos de los boreholes para leerlos
document_name = df["Path"]
# for filename in os.listdir("../boreholes/raw"):
#     document_name.append(os.path.join("../boreholes/raw", filename))

# Función para calcular la cantidad de datos que tenemos, a partir de una profundidad
#===================================================================================
def mayor(depth:int,zj:list)->int:
    mayor = 0
    for l in zj:
        if(l>depth):
            mayor +=1
    return mayor


# Función para graficar los datos obtenidos de el archivo
#===================================================================================
def plot_data(Tj:list,zj:list,name:str)->plt.Figure:
    fig = plt.figure(figsize=(3, 5))
    plt.xlabel("Temperaturas (°C)")
    plt.ylabel("Profundidad (m)") 
    plt.gca().invert_yaxis()
    plt.title("Regresión lineal del borehole {}".format(name))


    plt.rcParams['xtick.bottom'] = plt.rcParams['xtick.labelbottom'] = False
    plt.rcParams['xtick.top'] = plt.rcParams['xtick.labeltop'] = True

    plt.plot (Tj,zj,"o",color = "black",markerfacecolor="none",markeredgecolor="black")
    plt.close(fig)
    
    return fig


# Función para calcular la regresión lineal
#=====================================================================================
def regresion(Tj:list, zj:list)-> Tuple[np.array, float]:
    # Me quedo con los datos en profundidades mayores a 100m
    zj_red = zj [len(zj)-mayor(150,zj):] 
    Tj_red = Tj [len(zj)-mayor(150,zj):]

    #Linear regression L2
    model = np.polyfit(zj_red,Tj_red,1)

    #The number 1 defines the degree of the polynomial you want to fit
    #Model stores the a and b values:
    #              y = a*x + b

    a = model[0] #dT/Dz [K/m]
    b = model[1]
    geothermal= a*1000

    T_steady = b + a*zj

    return T_steady, geothermal


# Función para graficar la regresión
#====================================================================================
def plot_regresion(T_steady:list,Tj:list,zj:list,name:str)->None:
    fig = plt.figure(figsize=(3, 5))
    plt.xlabel("Temperaturas (°C)")
    plt.ylabel("Profundidad (m)") 
    plt.gca().invert_yaxis()
    plt.title("Regresión lineal del borehole {}".format(name))


    plt.rcParams['xtick.bottom'] = plt.rcParams['xtick.labelbottom'] = False
    plt.rcParams['xtick.top'] = plt.rcParams['xtick.labeltop'] = True

    plt.plot (Tj,zj,"o",color = "black",markerfacecolor="none",markeredgecolor="black")
    plt.plot(T_steady,zj,"r")
    
    plt.savefig(f"../resultados/regresion/regresion_{name}.png")
    plt.close(fig)
    
    


# Plot de las anomalías
#=====================================================================================
def plot_anomalias(anom:list, zj:list, nombre:str)->None:
    fig = plt.figure(figsize=(3, 5))
    plt.xlim(-1,2)
    plt.ylim(0,500)

    plt.xlabel("Anomalías")
    plt.ylabel("Profundidad (m)") 
    plt.gca().invert_yaxis()

    plt.rcParams['xtick.bottom'] = plt.rcParams['xtick.labelbottom'] = False
    plt.rcParams['xtick.top'] = plt.rcParams['xtick.labeltop'] = True

    plt.plot (anom,zj,color = sns.color_palette('Set2')[2])
    plt.title("Anomalias de temperatura {}".format(nombre))
    
    plt.savefig("../resultados/anomalias/img/anomalias_{}.png".format(nombre))
    plt.close(fig)


# Defino una lista a la que le iremos añadiendo valores
geothermal = []

for i in range(len(names)):
    zj,Tj = np.loadtxt (document_name[i],skiprows = int(posicion[i]+1),unpack = True,encoding="latin1")
    T_steady,geo = regresion(Tj=Tj,zj=zj)
    plot_regresion(T_steady,Tj,zj,names[i])
    
    #Caculo anomalías
    Tt = Tj-T_steady

    #Guardo el valor de las anomalías
    np.savetxt(f"../resultados/anomalias/csv/anomalias_{names[i]}.csv",Tt,delimiter=",")

    #Hago su gráfico
    plot_anomalias(Tt,zj,names[i])

    #Guardo el valor del gradiente geotérmico
    geothermal.append(geo)

# Los gráficos de las regresiones se guardan en la ruta: australia\resultados\regresion\
# Los gráficos de las anomalias se guardan en la ruta:   australia\resultados\anomalias\img\

# Añado la columna del gradiente geotérmico estimado al el dataframe.
df["Gradiente Geotérmico Experimental (K/km)"] = geothermal
df.head(1)
