==> australia/programas/04_inverse_problem.ipynb <==
## 04 - Problema Inverso

La inversión del perfil de temperaturas en un *borehole*, es una operación que transforma un perfil temperatura vs. profundidad en un tiempo determinado, a un perfil de temperatura vs. tiempo en una profundidad específica (la superficie). Y se hará a través del método llamado Descomposición en Valores Singulares (*Singular Value Decomposition*, SVD).

Problema directo:

$\Theta_j = T_k \thinspace A_{jk}$

Problema inverso con SVD:

$T_k =  A_{jk}^{-1} \Theta_j  = \mathbf{V}\thinspace\mathbf{\Lambda_p}^{-1}\thinspace\mathbf{U_p}^T\Theta_j$

# Importar librerías de tratamiento de datos
import numpy as np
import pandas as pd

# Visualización de datos
import matplotlib.pyplot as plt
import seaborn as sns
sns.set_theme(style="whitegrid")

# Otras
import warnings
from typing import Tuple
warnings.filterwarnings("ignore")

plt.ioff()

# Leo los datos de los 57 boreholes.
df = pd.read_csv("../boreholes/dataframe.csv",index_col="Unnamed: 0")

# Listas que nos van a ser últiles
names = list(df.index)
datos_menor40=np.genfromtxt("../resultados/datos/menor_40m.csv",delimiter=",",dtype="int")

# Defino varios parámetros 
#num_lambdas = 3
tk=list(range(20,501,20))


# Función para contar la cantidad la lambdas que se necesitan.
#===================================================================
def cantidad_lambdas(sigma:list)->int:
    suma=0
    for i in sigma:
        if(i>0.05):
            suma=suma+1
    return suma


# Función para calcular el problema inverso.
#=======================================================================
def inverse_problem(Ail:np.array,Tt:list)->Tuple[np.array, np.array]:
    # Modelo de singular value decomposition
    U,sigma,Vt = np.linalg.svd(Ail)

    V = np.transpose(Vt)
    Ut = np.transpose(U)

    num_lambdas = cantidad_lambdas(sigma)
    sigma_red = sigma[:3] 

    Ut_red1 = Ut[0:len(sigma_red)]
    U_red= np.transpose(Ut_red1)
    Vt_red1 = Vt[0:len(sigma_red)]
    V_red = np.transpose(Vt_red1) 

    diagonal_sigma_red_inv = np.diag(1/np.array(sigma_red))

    mest = V_red @ (Ut_red1 @ Tt @ diagonal_sigma_red_inv)
    
    return mest,sigma


# Plot de los valores propios
#========================================================================
def plot_lambdas(sigma:np.array,nombre:str)->None:
    index = range(len(sigma))
    
    fig = plt.figure(i)
    plt.plot(sigma,index,"o")
    plt.ylabel("index,i")
    plt.xlabel("lambda, $\lambda$")
    plt.title("Valores Singulares de la Matriz {}.".format(names[i]))
    plt.savefig(f"../resultados/mest/graficos/lambdas_{nombre}")
    plt.close(fig)

# Defino un dataframe vacio.
df_mest = pd.DataFrame(columns = names,index=tk[1:])

# Cálculo de las anomalías estimadas en la superficie de la Tierra.
for i in range(len(names)):

    # Leo las matrices de los diferentes boreholes
    df=pd.read_csv(f"../resultados/matriz/matriz_{names[i]}.csv")
    Ail = df.to_numpy()

    # Leo las anomalías en diferentes profundidades
    Tt = np.genfromtxt(f"../resultados/anomalias/csv/anomalias_{names[i]}.csv",delimiter=",")
    Tt = Tt[datos_menor40[i]:]

    mest,sigma=inverse_problem(Ail,Tt)
    df_mest ["{}".format(names[i])] = mest
    mest = list(mest)

    #Guardo los valores mest
    np.savetxt(f"../resultados/mest/csv/mest_{names[i]}.csv",Tt,delimiter=",")

    plot_lambdas(sigma,names[i])

# Los gráficos de los valores singulares se guardan en la ruta: australia\resultados\mest\graficos\

# Guardo los vectores estimados en un dataframe.
df_mest.to_csv("../resultados/mest/mest.csv",index=None)
df_mest.head()
